+++ packages/block-library/src/description-detail/save.js
@@ -1,0 +1,9 @@
+import { RichText, useBlockProps } from '@wordpress/block-editor';
+
+export default function save( { attributes } ) {
+	return (
+		<dd { ...useBlockProps.save() }>
+			<RichText.Content value={ attributes.content } />
+		</dd>
+	);
+}