+++ packages/block-library/src/description-list/save.js
@@ -1,0 +1,5 @@
+import { useBlockProps, useInnerBlocksProps } from '@wordpress/block-editor';
+
+export default function save() {
+	return <dl { ...useInnerBlocksProps.save( useBlockProps.save() ) } />;
+}